requirements.txt
streamlit==1.42.2
langchain==0.3.20
langchain-google-genai==2.1.0
python-dotenv==1.1.0

